+++ developer/tester.py
@@ -1,0 +1,4 @@
+from os import chdir, system
+
+chdir('..')
+system('rake robot')